reduce duplicate GitHub Actions runs
Since GitHub Actions workflows are triggered on every push,
triggering on pull request is usually redundant.
The exception is when pull requests are opened across repositories.
However, it should be sufficient to trigger a build on pull request
if the target branch is 'master'. This will ensure that changes
can't be merged into the master branch without being tested first,
but will not always trigger extra builds for pull requests internal
to a repo.